update billing notice
